Fix widemode resolution bug and change bestfit to a viewmode
Previously, hitting 'w' to fit the image to the width of the screen
would show an unreasonably low resolution version of images. That is
now fixed.

Also, "best fit" (auto rotate landscape/portrait to use the most
screen space) is now a viewmode, the same as "full", "fast", and
"width".