Use awk when untrashing files
Need to get the filename from a file that looks like:

	[Trash Info]
	Path=/home/ben/myprojects/vv/f(o)o%5C'&$%7Bx%7D%3C%7C%3E!.gif
	DeletionDate=Y-05-20T15:44:21

Previously was doing `eval $(grep Path filename.info)` to set Path,
but that was silly as certain metacharacters that are special to the
shell are not special to URIs and aren't percent encoded. [Okay, to be
honest, using "eval" is always silly.]

Fixed it by simply using awk instead of grep so I could output just
the filename. That replaces a dozen lines of kludges to make it work
(+ more lines of comments for explanations) with a single line. Yay!